Removed timeout feature, renamed query funcs.
Removed timeout option. May re-implement later after 1.0 release if
someone wants it. I want to implement it in a thread instead of in loops
in different parts of main().

Renamed get_* functions to query_* functions if they use query_api().
Easier to track what hits the API and what doesn't.